+++ src/test/kotlin/domainTest/gameTest/CarTest.kt
@@ -1,4 +1,4 @@
-package domaintest
+package domainTest.gameTest
 
 import org.assertj.core.api.Java6Assertions.assertThat
 import org.junit.jupiter.api.Test
